### 2025-07-04 #8222 @abhishektiwari
@DingoGoesBingo Can you action following before we can commence search for reviewers?

- [ ] Please add guidelines for contribution `CONTRIBUTING.md`. See [a generator here](https://contributing.md/generator/).
- [ ] Remove the hardcoded references in `paper.md` after `# References`. References will be injected using the `paper.bib`
- [ ] Add missing DOI for references as highlighted by editorial bot
- [ ] Can you also clarify the co-author's contribution to software and paper? I do not see any [code contribution](https://github.com/DingoGoesBingo/Seed-Database-Manager/graphs/contributors) from your co-author.

```
- 10.32614/cran.package.rpostgres may be a valid DOI for title: RPostgres: C++ Interface to PostgreSQL
- 10.1093/database/baab054 may be a valid DOI for title: The Breeding Information Management System (BIMS):...
- 10.1007/978-1-59745-535-0_22 may be a valid DOI for title: International crop information system for germplas...
- 10.2135/cropsci2016.09.0814 may be a valid DOI for title: Germinate 3: development of a common platform to s...
- 10.1007/978-981-19-5434-4_2 may be a valid DOI for title: Germplasm management in commercial plant breeding ...

```
