### 2025-07-08 e21bb78
Added templates for issues and feature requests
