+++ .github/ISSUE_TEMPLATE/feature-request-form.md
@@ -1,0 +1,20 @@
+---
+name: Feature request form
+about: Propose new features for this project
+title: ''
+labels: ''
+assignees: DingoGoesBingo
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is.
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen, and what new features could be added to alleviate issues.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
