+++ .github/ISSUE_TEMPLATE/bug-report-form.md
@@ -1,0 +1,36 @@
+---
+name: Bug report form
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: DingoGoesBingo
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Please do your best to describe the steps taken to produce this bug, if it is reproducible:
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**For desktop/computers**
+Please provide the following information: 
+- OS [e.g. Windows 11, MacOS, etc]
+ - Browser [e.g. chrome, safari]
+ - Version [e.g. 22]
+
+**For smartphone (if using online app)**
+Please provide the following information:
+ - Device: [e.g. iPhone6]
+ - OS: [e.g. iOS8.1]
+ - Browser [e.g. stock browser, safari]
+ - Version [e.g. 22]
+
+**Additional context**
+Add any other context about the problem here.
