### 2025-07-08 f43e462
Updated paper.bib, paper.md, added contibution guidelines & code of conduct
