+++ paper.bib
@@ -4,6 +4,7 @@
   year = {2024},
   note = {R package version 1.4.7},
   url = {https://CRAN.R-project.org/package=RPostgres},
+  doi = {10.32614/CRAN.package.RPostgres}
 }
 
 @article{jung2021breeding:2021,
@@ -14,6 +15,7 @@
   pages={baab054},
   year={2021},
   publisher={Oxford University Press UK}
+  doi = {10.1093/database/baab054}
 }
 
 @article{morales2022breedbase:2022,
@@ -25,6 +27,7 @@
   pages={jkac078},
   year={2022},
   publisher={Oxford University Press}
+  doi = {10.1093/g3journal/jkac078}
 }
 
 @article{portugal2007international:2007,
@@ -34,6 +37,7 @@
   pages={459--471},
   year={2007},
   publisher={Springer}
+  doi = {10.1007/978-1-59745-535-0_22}
 }
 
 @article{shaw2017germinate:2017,
@@ -45,6 +49,7 @@
   pages={1259--1273},
   year={2017},
   publisher={Wiley Online Library}
+  doi = {10.2135/cropsci2016.09.0814}
 }
 
 @incollection{tiwari2023germplasm:2023,
@@ -54,6 +59,7 @@
   pages={33--68},
   year={2023},
   publisher={Springer}
+  doi = {10.1007/978-981-19-5434-4_2}
 }
 
 @Manual{shiny:2024,
@@ -62,4 +68,5 @@
   year = {2024},
   note = {R package version 1.8.1.1},
   url = {https://CRAN.R-project.org/package=shiny},
+  doi = {https://doi.org/10.32614/CRAN.package.shiny}
 }
