### 2025-07-09 84d7a59
Fix paper.bib validation failure
Fix paper.bib validation failure
