+++ paper.bib
@@ -14,7 +14,7 @@
   volume={2021},
   pages={baab054},
   year={2021},
-  publisher={Oxford University Press UK}
+  publisher={Oxford University Press UK},
   doi = {10.1093/database/baab054}
 }
 
@@ -26,7 +26,7 @@
   number={7},
   pages={jkac078},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi = {10.1093/g3journal/jkac078}
 }
 
@@ -36,7 +36,7 @@
   journal={Plant Bioinformatics: Methods and Protocols},
   pages={459--471},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/978-1-59745-535-0_22}
 }
 
@@ -48,7 +48,7 @@
   number={3},
   pages={1259--1273},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi = {10.2135/cropsci2016.09.0814}
 }
 
@@ -58,7 +58,7 @@
   booktitle={Market-Driven Plant Breeding for Practicing Breeders},
   pages={33--68},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/978-981-19-5434-4_2}
 }
 
