### 2025-07-09 #8222 @abhishektiwari
@DingoGoesBingo Your `paper.bib` updates are failing validation. When you added you `doi` you missed to add `,` on previous line. So

```
@article{jung2021breeding:2021,
  title={The Breeding Information Management System (BIMS): an online resource for crop breeding},
  author={Jung, Sook and Lee, Taein and Gasic, Ksenija and Campbell, B Todd and Yu, Jing and Humann, Jodi and Ru, Sushan and Edge-Garza, Daniel and Hough, Heidi and Main, Dorrie},
  journal={Database},
  volume={2021},
  pages={baab054},
  year={2021},
  publisher={Oxford University Press UK}
  doi = {10.1093/database/baab054}
}

```

Should be

```
@article{jung2021breeding:2021,
  title={The Breeding Information Management System (BIMS): an online resource for crop breeding},
  author={Jung, Sook and Lee, Taein and Gasic, Ksenija and Campbell, B Todd and Yu, Jing and Humann, Jodi and Ru, Sushan and Edge-Garza, Daniel and Hough, Heidi and Main, Dorrie},
  journal={Database},
  volume={2021},
  pages={baab054},
  year={2021},
  publisher={Oxford University Press UK},
  doi = {10.1093/database/baab054}
}
```

Can you please review [this CR and merge it](https://github.com/DingoGoesBingo/Seed-Database-Manager/pull/1) and then try to run `@editorialbot generate pdf` to make sure validation is passing and PDF is being rendered.
