### 2025-07-09 #8222 @DingoGoesBingo
Hello! We could potentially reach out to Francisco Agosto (GitHub username @/agostof) at Cornell University, since he does similar database work (mostly in Python, but some projects in R).
