### 2025-07-12 #8222 @abhishektiwari
👋 @agostof @xoolive @algarsi3, would any of you be willing to review this submission for JOSS?  The submission under consideration is "Seed Database Manager: an open-source tool for germplasm information management".

Based on your experience, we think you might be able to provide a great review of this submission.

Thanks
Abhi
