### 2025-08-05 #8222 @DingoGoesBingo
Hello! I reached out to a colleague of mine and they were able to share a few names, here is a short list:

Anna McCormick [ahmccorm@hawaii.edu]
Aaron Lorenz [lore0149@umn.edu]
Nathan Fumia [nfumia@harc-hspa.com]
Yue Yu [yue.yu@ubc.ca]
Maarten van Zonneveld [maarten.vanzonneveld@worldveg.org]

I already personally emailed Aaron Lorenz and haven't received a response. Hope this helps! I'll continue asking around as well.
