### 2025-08-08 #8222 @abhishektiwari
👋  @DS4Ag @RobLBaker @nehakn, would any of you be willing to review this submission for JOSS?  The submission under consideration is "Seed Database Manager: an open-source tool for germplasm information management".

Based on your experience and your research interest, I think you might be able to provide a good review of this submission.

Thanks
Abhi
