### 2025-08-19 #8835 @editorialbot
[REVIEW]: Seed Database Manager: an open-source tool for germplasm information management
**Submitting author:** @DingoGoesBingo (<a href="http://orcid.org/0009-0002-8034-6783">Sam Schafer</a>)
**Repository:** https://github.com/DingoGoesBingo/Seed-Database-Manager
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
