### 2025-09-19 #8835 @jsun
Hi, @DingoGoesBingo

Here are my preliminary comments; I will provide a full report after completing the review.

- The installation instructions (README.md in the repository root) are difficult to follow. I was unsure how to proceed when my Railway dashboard displayed “build failed” at this step. I eventually realized that I needed to click the “redeploy” button, but then encountered errors during the Docker image build process `[8/19] RUN R -e "install.packages(c('shiny', 'shinyjs', 'shinythemes', 'shinyalert', 'stringr','ggplot2', 'DT', 'officer', 'base64enc'))"`.
- I also checked https://sdm.up.railway.app/
, mentioned in the Manual directory, and experimented with the app. Does this app require Postgres? If it could instead support SQLite or even just use a CSV file, the deployment process would be greatly simplified.

To continue the review, could you please verify whether the installation can be completed successfully? I tried this with Railway. Thank you.
