### 2025-10-27 #8835 @DingoGoesBingo
> I retried installing the application in a local environment and was finally able to install it successfully. However, the installation process was quite difficult and could be improved for usability.
>
> **manual**
>
> * Overall, the current version of the manual is difficult to follow. I recommend that the developers create documentation using GitHub Pages or a Wiki, which would simplify the installation instructions. Adding screenshots or diagrams would also help explain the steps more clearly.
> * In Step 1, I installed PostgreSQL on macOS, but it did not include pgAdmin. The manual should note that users need to install both PostgreSQL and pgAdmin.
> * In Step 5, when running the R script in RStudio, I found that the working directory needs to be set to the Setup folder. This important detail is not currently mentioned in the manual.
>
> **app**
>
> * The application appears to be designed for small-scale teams, but the overall structure — from installation to source code — feels somewhat complex. For example, the seed data mainly consist of text and numerical values, which could easily be handled using SQLite. However, the current version only supports PostgreSQL, which increases complexity and makes installation more difficult. Additionally, configuring database parameters (ID, password, host, port) can be challenging for small-scale users. Therefore, I strongly recommend that the developers add support for SQLite to simplify deployment.
> * In Step 6.1, adding data currently requires direct manipulation of PostgreSQL tables. The application should instead provide a user interface for data input, avoiding the need for direct database operations.
>
> To enhance usability, I recommend that the developers provide one of the following options:
>
> * A Docker image that includes all necessary configurations, allowing users to access the application immediately after running the container.
> * A self-contained Shiny application (using SQLite), enabling users to easily install and launch the app without complex setup steps.

Hello! Sorry I've been slow to reply, I have been really busy teaching and simultaneously prepping for an upcoming conference -- I promise I'm not ignoring these comments!

I will work on improving the clarity of the github instructions soon and work on including a pre-made docker image to include for local setup. The self-contained suggestion is a good idea, my original idea was to make this an application that you just install but I hadn't found a method to do so just yet -- so I'll look into SQLite!

It may be a few weeks before I can address all of these, so sorry in advance!
