### 2025-11-10 #8835 @abhishektiwari
@DingoGoesBingo As a courtesy to the reviewers and in keeping with our JOSS goal of generally completing a review within 4 to 6 weeks, can you aim to tackle installation related issue in next 2 weeks.  We are coming to now 8 weeks and reviewers have been unable to install the software due to various issues and we have hardly made any progress.

If you are unable to do so and given current state of software, I will strongly suggest to withdraw your submission and re-submit it to JOSS when installation and usability issues highlighted by reviewers are tackled.
